% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{fit_av}
\alias{fit_av}
\title{Computes the Allan Variance}
\usage{
fit_av(x, qn = NULL, wn = NULL, rw = NULL, dr = NULL, type = NULL)
}
\arguments{
\item{x}{A \code{vec} of time series observations or an \code{avar} object.}

\item{qn}{A \code{vec} specifying on which scales the parameters of a Quantization Noise (QN) should be computed.}

\item{wn}{A \code{vec} specifying on which scales the parameters of a White Noise (WN) should be computed.}

\item{rw}{A \code{vec} specifying on which scales the parameters of a Random Wakk (RW) should be computed.}

\item{dr}{A \code{vec} specifying on which scales the parameters of a Drift (DR) should be computed.}

\item{type}{A \code{string} containing either \code{"mo"} for Maximal Overlap or \code{"to"} for Tau Overlap}
}
\value{
Fit_av   A \code{list} that contains:
\itemize{
 \item{"process"}{The process selected}
 \item{"theta_hat"}{The estimated parameter vector}
}
}
\description{
Compute the latent processes parameters estimator based on the Allan Variance
}
\examples{
# Load simts package
library(simts)

# Set seed for reproducibility
set.seed(999)

# Simulate time series
N = 100000
x = gen_gts(N, WN(sigma2 = 2) + RW(gamma2 = 1))

# Maximal overlap
fit1 = fit_av(x, wn = 1:10, rw = 11:14)

}
