% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avlr.R
\name{fit_avlr}
\alias{fit_avlr}
\title{Internal function to the Allan Variance Linear Regression estimator}
\usage{
fit_avlr(qn, wn, rw, dr, ad, scales)
}
\arguments{
\item{qn}{A \code{vec} specifying on which scales the parameters of a Quantization Noise (QN) should be computed.}

\item{wn}{A \code{vec} specifying on which scales the parameters of a White Noise (WN) should be computed.}

\item{rw}{A \code{vec} specifying on which scales the parameters of a Random Wakk (RW) should be computed.}

\item{dr}{A \code{vec} specifying on which scales the parameters of a Drift (DR) should be computed.}

\item{ad}{A \code{vec} of the Allan variance.}

\item{scales}{A \code{vec} of the scales.}
}
\value{
A \code{list} with the estimated parameters.
}
\description{
Estimate the parameters of time series models based on the Allan Variance Linear Regression (AVLR) approach
}
