% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imar_av}
\alias{imar_av}
\title{#' @title Allan variance of IMU Data from IMAR Gyroscopes
#' @description This data set contains Allan variance of IMAR gyroscopes data.
#' @format A list of the following elements:
#' \describe{
#'  \item{\code{sensor}}{Name of the sensor.}
#'  \item{\code{freq}}{The frequency at which the error signal is measured.}
#'  \item{\code{n}}{Sample size of the data.}
#'  \item{\code{type}}{The types of sensors considered in the data.}
#'  \item{\code{axis}}{The axes of sensors considered in the data.}
#'  \item{\code{avar}}{A list containing the computed Allan variance based on the data.}
#' }
#' @source The IMU data comes from Geodetic Engineering Laboratory (TOPO) and Swiss Federal Institute of Technology Lausanne (EPFL).}
\format{An object of class \code{imu_avar} of length 6.}
\usage{
imar_av
}
\description{
#' @title Allan variance of IMU Data from IMAR Gyroscopes
#' @description This data set contains Allan variance of IMAR gyroscopes data.
#' @format A list of the following elements:
#' \describe{
#'  \item{\code{sensor}}{Name of the sensor.}
#'  \item{\code{freq}}{The frequency at which the error signal is measured.}
#'  \item{\code{n}}{Sample size of the data.}
#'  \item{\code{type}}{The types of sensors considered in the data.}
#'  \item{\code{axis}}{The axes of sensors considered in the data.}
#'  \item{\code{avar}}{A list containing the computed Allan variance based on the data.}
#' }
#' @source The IMU data comes from Geodetic Engineering Laboratory (TOPO) and Swiss Federal Institute of Technology Lausanne (EPFL).
}
\keyword{datasets}
