% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{navchip_av}
\alias{navchip_av}
\title{Allan variance of IMU Data from a navchip sensor}
\format{A list of the following elements:
\itemize{
 \item "sensor": Name of the sensor.
 \item "freq": The frequency at which the error signal is measured.
 \item "n": Sample size of the data.
 \item "type": The types of sensors considered in the data.
 \item "axis": The axes of sensors considered in the data.
 \item "avar": A list containing the computed Allan variance based on the data.
}}
\source{
The IMU data of the navchip sensor comes from Geodetic Engineering Laboratory (TOPO) and Swiss Federal Institute of Technology Lausanne (EPFL).
}
\usage{
navchip_av
}
\description{
This data set contains Allan variance of gyroscope and accelerometer data from a navchip sensor.
}
\keyword{datasets}
