% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avlr.R
\name{plot.avlr}
\alias{plot.avlr}
\title{Plot the AVLR with the Allan Deviation}
\usage{
\method{plot}{avlr}(
  x,
  decomp = FALSE,
  units = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col_ad = NULL,
  col_ci = NULL,
  nb_ticks_x = NULL,
  nb_ticks_y = NULL,
  legend_position = NULL,
  ci_ad = NULL,
  point_cex = NULL,
  point_pch = NULL,
  show_scales = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{avlr} object.}

\item{decomp}{A \code{boolean} that determines whether the contributions of each individual model are plotted.}

\item{units}{A \code{string} that specifies the units of time plotted on the x axis.}

\item{xlab}{A \code{string} that gives a title for the x axis.}

\item{ylab}{A \code{string} that gives a title for the y axis.}

\item{main}{A \code{string} that gives an overall title for the plot.}

\item{col_ad}{A \code{string} that specifies the color of the line allan deviation line.}

\item{col_ci}{A \code{string} that specifies the color of the shaded area covered by the confidence intervals.}

\item{nb_ticks_x}{An \code{integer} that specifies the maximum number of ticks for the x-axis.}

\item{nb_ticks_y}{An \code{integer} that specifies the maximum number of ticks for the y-axis.}

\item{legend_position}{A \code{string} that specifies the position of the legend (use \code{legend_position = NA} to remove legend).}

\item{ci_ad}{A \code{boolean} that determines whether to plot the confidence interval shaded area.}

\item{point_cex}{A \code{double} that specifies the size of each symbol to be plotted.}

\item{point_pch}{A \code{double} that specifies the symbol type to be plotted.}

\item{show_scales}{A \code{boolean} that specifies if the scales used for each process should be plotted.}

\item{...}{Additional arguments affecting the plot.}
}
\value{
Plot of Allan deviation and relative confidence intervals for each scale.
}
\description{
Displays a plot of the Allan deviation (AD) with the CI values and the AD implied by the estimated parameters.
}
\examples{
\donttest{
set.seed(999)

N = 100000
Xt = rnorm(N) + cumsum(rnorm(N, 0, 3e-3))
av = avlr(Xt, wn = 1:7, rw = 12:15)

plot.avlr(av)
plot.avlr(av, decomp = TRUE, main = "Simulated white noise", xlab = "Scales")
plot.avlr(av, units = "sec", legend_position = "topright")
plot.avlr(av, col_ad = "darkred", col_ci = "pink")
plot(fit, decomp = TRUE, show_scales = TRUE)
}
}
\author{
Stephane Guerrier and Justin Lee
}
