% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{plot.imu_avlr}
\alias{plot.imu_avlr}
\title{TO DO}
\usage{
\method{plot}{imu_avlr}(x, xlab = NULL, ylab = NULL, main = NULL,
  col_ad = NULL, col_ci = NULL, nb_ticks_x = NULL,
  nb_ticks_y = NULL, ci_ad = NULL, point_pch = NULL,
  point_cex = NULL, ...)
}
\description{
TO DO
}
