% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{print.avar}
\alias{print.avar}
\title{Prints Allan Variance}
\usage{
\method{print}{avar}(x, ...)
}
\arguments{
\item{x}{A \code{avar} object.}

\item{...}{Arguments to be passed to methods}
}
\value{
console output
}
\description{
Displays the information on the output of the `avar()` function
}
\examples{
# Set seed for reproducibility
set.seed(999)

# Simulate a Gaussian white noise
Xt = rnorm(10000)

# Compute Allan
out = avar(Xt)

# Print results
print(out)

}
