% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{print.imu_avlr}
\alias{print.imu_avlr}
\title{Print imu_avlr object}
\usage{
\method{print}{imu_avlr}(x, ...)
}
\arguments{
\item{x}{A \code{imu_avlr} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays information about the imu_avlr object
}
\examples{
\donttest{
# TO DO
}
}
\keyword{internal}
