% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{avlr}
\alias{avlr}
\title{Computes the Allan Variance Linear Regression estimator}
\usage{
avlr(x, qn = NULL, wn = NULL, rw = NULL, dr = NULL, ci = FALSE,
  B = 100, alpha = 0.05)
}
\arguments{
\item{x}{A \code{vec} of time series observations or an \code{imu} object.}

\item{qn}{A \code{vec} specifying on which scales the parameters of a Quantization Noise (QN) should be computed.}

\item{wn}{A \code{vec} specifying on which scales the parameters of a White Noise (WN) should be computed.}

\item{rw}{A \code{vec} specifying on which scales the parameters of a Random Wakk (RW) should be computed.}

\item{dr}{A \code{vec} specifying on which scales the parameters of a Drift (DR) should be computed.}

\item{ci}{A \code{boolean} to compute parameter confidence intervals.}

\item{B}{A \code{double} for the number of bootstrap replicates to compute the parameter confidence intervals.}

\item{alpha}{A \code{double} defining the level of the confidence interval (1 - `alpha`).}
}
\value{
avlr   A \code{list} that contains:
\itemize{
 \item{"estimates"}{The estimated value of the parameters.}
 \item{"implied_ad"}{The Allan deviation implied by the estimated parameters.}
 \item{"implied_ad_decomp"}{The Allan deviation implied by the estimated parameters for each individual model (if more than one is specified).}
 \item{"av"}{The \code{avar} object computed from the provided data.}
}
}
\description{
Estimate the parameters of time series models based on the Allan Variance Linear Regression (AVLR) approach
}
\examples{
# Load simts package
library(simts)

# Set seed for reproducibility
set.seed(99)

# Simulate time series

N = 100000
x = gen_gts(N, WN(sigma2 = 1)  + RW(gamma2 = 1e-5))

# Compute Allan variance
av = avar(x)
plot(av)

# Parameter estimation
fit = avlr(x, wn = 1:8, rw = 11:15)
plot(fit, decomp = TRUE)

# Point estimates
fit

# Compute confidence intervals (this step is time-demanding)
fit = avlr(x, wn = 1:8, rw = 10:15, ci = TRUE, B = 30)

# Estimated confidence intervals and standard deviations
fit$ci
}
