% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{avlr}
\alias{avlr}
\title{Computes the Allan Variance Linear Regression estimator}
\usage{
avlr(x, qn = NULL, wn = NULL, rw = NULL, dr = NULL, type = NULL)
}
\arguments{
\item{x}{A \code{vec} of time series observations or an \code{avar} object.}

\item{qn}{A \code{vec} specifying on which scales the parameters of a Quantization Noise (QN) should be computed.}

\item{wn}{A \code{vec} specifying on which scales the parameters of a White Noise (WN) should be computed.}

\item{rw}{A \code{vec} specifying on which scales the parameters of a Random Wakk (RW) should be computed.}

\item{dr}{A \code{vec} specifying on which scales the parameters of a Drift (DR) should be computed.}

\item{type}{A \code{string} containing either \code{"mo"} for Maximal Overlap or \code{"to"} for Tau Overlap}
}
\value{
avlr   A \code{list} that contains:
\itemize{
 \item{"estimates"}
 \item{"implied_ad"}{The Allan Deviation implied by the estimated parameter.}
 \item{"implied_ad_decomp"}{The Allan Deviation implied by the estimated parameter for the sub-processes.}
 \item{"av"}{The \code{avar} object provided or corresponding to the data provided.}
}
}
\description{
Compute the latent processes parameters estimator based on the Allan Variance
}
\examples{
# Load simts package
library(simts)

# Set seed for reproducibility
set.seed(999)

# Simulate time series
N = 100000
x = gen_gts(N, WN(sigma2 = 1) + RW(gamma2 = 1e-7))

# Maximal overlap
fit1 = avlr(x, wn = 1:12, rw = 12:15)

}
