% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{avlr}
\alias{avlr}
\title{Computes the Allan Variance Linear Regression estimator}
\usage{
avlr(x, qn = NULL, wn = NULL, rw = NULL, dr = NULL, type = "mo",
  ci = FALSE, B = 100, alpha = 0.05)
}
\arguments{
\item{x}{A \code{vec} of time series observations, an \code{imu} object or an \code{avar} object.}

\item{qn}{A \code{vec} specifying on which scales the parameters of a Quantization Noise (QN) should be computed.}

\item{wn}{A \code{vec} specifying on which scales the parameters of a White Noise (WN) should be computed.}

\item{rw}{A \code{vec} specifying on which scales the parameters of a Random Wakk (RW) should be computed.}

\item{dr}{A \code{vec} specifying on which scales the parameters of a Drift (DR) should be computed.}

\item{type}{A \code{string} containing either \code{"mo"} (default) for Maximal Overlap or \code{"to"} for Tau Overlap}

\item{ci}{A \code{bolean} to compute confidence intervals for the parameter.}

\item{B}{A \code{double} for the number of bootsrap replicates to compute the confidence intervals.}

\item{alpha}{A \code{double} for the level of confidence \code{alpha} to compute the confidence intervals.}
}
\value{
avlr   A \code{list} that contains:
\itemize{
 \item{"estimates"}
 \item{"implied_ad"}{The Allan Deviation implied by the estimated parameter.}
 \item{"implied_ad_decomp"}{The Allan Deviation implied by the estimated parameter for the sub-processes.}
 \item{"av"}{The \code{avar} object provided or corresponding to the data provided.}
}
}
\description{
Estimate the parameters of time series models based on the Allan Variance Linear Regression (AVLR) approach
}
\examples{
# Load simts package
library(simts)

# Set seed for reproducibility
set.seed(99)

# Simulate time series

N = 100000
x = gen_gts(N, WN(sigma2 = 1)  + RW(gamma2 = 1e-5))

# Compute av
av = avar(x)
plot(av)

# Parameter estimation
fit = avlr(x, wn = 1:8, rw = 11:15)
plot(fit, decomp = TRUE)

# Point estimates
fit

# Compute confidence intervals (this can take some time)
fit = avlr(av, wn = 1:8, rw = 10:15, ci = TRUE, B = 30)

# Estimated confidence intervals and standard deviations
fit$ci
}
