% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avlr.R
\name{print.avlr}
\alias{print.avlr}
\title{Print avlr object}
\usage{
\method{print}{avlr}(x, ...)
}
\arguments{
\item{x}{A \code{avlr} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays information about the avlr object
}
\examples{
\donttest{
set.seed(999)

N = 100000
Xt = rnorm(N) + cumsum(rnorm(N, 0, 3e-3))

fit = avlr(Xt, wn = 1:7, rw = 12:15)
print(fit)
}
}
\keyword{internal}
