% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{print.avlr}
\alias{print.avlr}
\title{TO DO}
\usage{
\method{print}{avlr}(x, ...)
}
\arguments{
\item{x}{A \code{avlr} object.}

\item{...}{Arguments to be passed to methods}
}
\value{
console output
}
\description{
TO DO + needs example
}
