% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avar.r
\name{print.avlr}
\alias{print.avlr}
\title{Print avlr object}
\usage{
\method{print}{avlr}(x, ...)
}
\arguments{
\item{x}{A \code{avlr} object}

\item{...}{Other arguments passed to specific methods}
}
\value{
Text output via print
}
\description{
Displays information about the avlr object
}
\examples{
\dontrun{
# Load simts package
library(simts)

# Set seed for reproducibility
set.seed(999)

# Simulate time series
N = 100000
x = gen_gts(N, WN(sigma2 = 1) + RW(gamma2 = 1e-7))

# Maximal overlap
fit1 = avlr(x, wn = 1:12, rw = 12:15)
print(fit1)
}
}
\keyword{internal}
